Move 'home' and 'search' into 'hiber' site (#12)
* Move 'home' and 'search' into 'hiber' site
* Move templates and static files to respective directories

We want to preserve the functionality from 'home' and 'search', but we
didn't want them scoped in the higher level directory. This change
simply moves them into 'hiber' without breaking the site.
