+++ hiber/settings/base.py
@@ -24,8 +24,8 @@ BASE_DIR = os.path.dirname(PROJECT_DIR)
 # Application definition
 
 INSTALLED_APPS = [
-    'home',
-    'search',
+    'hiber.home',
+    'hiber.search',
 
     'wagtail.contrib.forms',
     'wagtail.contrib.redirects',
